+++ src/stitch/execute.ts
@@ -6,6 +6,7 @@ import type {
   GraphQLSchema,
   IncrementalResult,
   OperationDefinitionNode,
+  VariableDefinitionNode,
 } from 'graphql';
 import {
   assertValidSchema,
@@ -21,14 +22,28 @@ import { isPromise } from '../predicates/isPromise.js';
 
 import { createRequest } from './createRequest.js';
 import { mapAsyncIterable } from './mapAsyncIterable.js';
-import type { ExecutionContext, Executor } from './Stitcher.js';
-import { Stitcher } from './Stitcher.js';
 
 export interface ExecutionArgs {
   schema: GraphQLSchema;
   document: DocumentNode;
   variableValues?: { readonly [variable: string]: unknown } | undefined;
   operationName?: string | undefined;
+  executor: Executor;
+}
+
+export type Executor = (args: {
+  document: DocumentNode;
+  variables?: { readonly [variable: string]: unknown } | undefined;
+}) => PromiseOrValue<ExecutionResult | ExperimentalIncrementalExecutionResults>;
+
+export interface ExecutionContext {
+  schema: GraphQLSchema;
+  fragments: Array<FragmentDefinitionNode>;
+  fragmentMap: ObjMap<FragmentDefinitionNode>;
+  operation: OperationDefinitionNode;
+  variableDefinitions: ReadonlyArray<VariableDefinitionNode>;
+  rawVariableValues: { readonly [variable: string]: unknown } | undefined;
+  coercedVariableValues: { [variable: string]: unknown };
   executor: Executor;
 }
 
@@ -47,11 +62,9 @@ export function execute(
   const result = delegate(exeContext);
 
   if (isPromise(result)) {
-    return result.then((resolved) =>
-      handlePossibleMultiPartResult(exeContext, resolved),
-    );
+    return result.then((resolved) => handlePossibleMultiPartResult(resolved));
   }
-  return handlePossibleMultiPartResult(exeContext, result);
+  return handlePossibleMultiPartResult(result);
 }
 
 export function buildExecutionContext(
@@ -158,10 +171,10 @@ function delegate(
 
 function handlePossibleMultiPartResult<
   T extends ExecutionResult | ExperimentalIncrementalExecutionResults,
->(exeContext: ExecutionContext, result: T): PromiseOrValue<T> {
+>(result: T): PromiseOrValue<T> {
   if ('initialResult' in result) {
     return {
-      initialResult: new Stitcher(exeContext, result.initialResult).stitch(),
+      initialResult: result.initialResult,
       subsequentResults: mapAsyncIterable(
         result.subsequentResults,
         (payload) => {
@@ -171,7 +184,7 @@ function handlePossibleMultiPartResult<
 
             let containsPromises = false;
             for (const entry of payload.incremental) {
-              const stitchedEntry = new Stitcher(exeContext, entry).stitch();
+              const stitchedEntry = entry;
               if (isPromise(stitchedEntry)) {
                 containsPromises = true;
               }
@@ -191,5 +204,5 @@ function handlePossibleMultiPartResult<
     } as T;
   }
 
-  return new Stitcher(exeContext, result).stitch() as T;
+  return result;
 }
